# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/k8sPods/createK8sPods": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8sPods"
                ],
                "summary": "创建K8sPods",
                "parameters": [
                    {
                        "description": "创建K8sPods",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.K8sPods"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/k8sPods/deleteK8sPods": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8sPods"
                ],
                "summary": "删除K8sPods",
                "parameters": [
                    {
                        "description": "删除K8sPods",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.K8sPods"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/k8sPods/deleteK8sPodsByIds": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8sPods"
                ],
                "summary": "批量删除K8sPods",
                "parameters": [
                    {
                        "description": "批量删除K8sPods",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"批量删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/k8sPods/execPod": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "调用后端接口http://127.0.0.1:8888/container/execContainer;打开前端web/src/view/kubernetes/container/terminal.html文件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8sPods"
                ],
                "summary": "进入容器终端",
                "parameters": [
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Pod名称",
                        "name": "podName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "containerName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"进入容器终端\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/k8sPods/getK8sPodsList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "默认获取集群ID为1的所有namespace的容器,获取指定集群的namespace的容器接口:/k8sPods/getK8sPodsList?clusterID={int}\u0026namespace={string}",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8sPods"
                ],
                "summary": "分页获取K8sPods列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "集群ID",
                        "name": "clusterID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/k8sPods/getPodLog": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8sPods"
                ],
                "summary": "获取pod日志",
                "parameters": [
                    {
                        "type": "string",
                        "default": "default",
                        "description": "命名空间",
                        "name": "namespace",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "podName",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "日志行数",
                        "name": "line",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取pod日志成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/k8sPods/updateK8sPods": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "K8sPods"
                ],
                "summary": "更新K8sPods",
                "parameters": [
                    {
                        "description": "更新K8sPods",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.K8sPods"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.K8sPods": {
            "type": "object",
            "properties": {
                "hostIP": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "line": {
                    "type": "integer"
                },
                "namespace": {
                    "type": "string"
                },
                "podIP": {
                    "type": "string"
                },
                "podName": {
                    "type": "string"
                },
                "restartCount": {
                    "type": "integer"
                },
                "startTime": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "request.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}